+++ tests/testthat/test-utils.R
@@ -50,13 +50,9 @@ test_that('REQUIRE TEST for combine_coef_se for bootstrapped', {
     expect_equal(round(as.numeric(
                     combine_coef_se(z.out.boot, bagging = TRUE)[[1]][1]), 3),
                  -0.052)
-})
 
-# FAIL TEST for combine_coef_se for non-mi or bootstrapped data ----------------
-test_that('FAIL TEST for combine_coef_se for non-mi or bootstrapped data', {
-    z5 <- zelig(Fertility ~ Education, data = swiss, model = 'ls')
-    expect_error(combine_coef_se(z5),
-                 'No multiply imputed or bootstrapped estimates found. So no need to combine.')
+    z5_ls <- zelig(Fertility ~ Education, model = "ls", data = swiss)
+    expect_equal(length(combine_coef_se(z5_ls)), 3)
 })
 
 # REQUIRE TEST for to_zelig_mi -------------------------------------------------